no ignore, update delegation command
