--- .gitignore
@@ -1,3 +1,0 @@
-stake-dist
-delegation
-
